Add sketch of operation count framework.
This is mainly a sketch of the interface to such a framework, illustrated only
for the SpMV kernel (because it is the one that dominates execution time). The
eaxmple API provides functions for tracing:

- Kernel start and end
- Loop iteration start and end and nesting
- Memory read
- Memory write

The implementation simply dumps everything to a trace file for now, but one
would imagine instead collecting the data and storing it in a form that can be
used to determine more interesting metrics for the memory accesses of functions
and experiment with simulating the trace on different devices.